incdec -> calc
